Added a native HTML loading message so that users with particularly slow browsers aren't stuck wondering what happened.
git-svn-id: http://bznetwork.googlecode.com/svn@537 dcf11152-862b-11de-82e5-a1837a67dc72

+++ trunk/bznetwork/main/src/jw/bznetwork/server/BZNetworkServer.java
@@ -1159,7 +1159,7 @@ public class BZNetworkServer implements ServletContextListener,
         {
             if (!Perms.server("view-logs", server, GlobalLinkImpl
                     .getServerGroupId(server)))
-                serverIds.remove(server);
+                serverIds.remove((Integer)server);
         }
         // add the servers
         filter += " and ( '1' = '2' ";
